requirements.txt
astroid==2.1.0
autopep8==1.4.3
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
dj-database-url==0.5.0
Django==2.2
django-crispy-forms==1.7.2
django-heroku==0.3.1
djangorestframework==3.9.2
gcloud==0.17.0
googleapis-common-protos==1.5.9
gunicorn==19.9.0
httplib2==0.12.1
idna==2.8
isort==4.3.4
jws==0.1.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauth2client==3.0.0
ParallelDots==3.2.12
Pillow==6.0.0
pipenv==2018.11.26
protobuf==3.7.1
psycopg2==2.8.2
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.5.0
pycryptodome==3.4.3
pylint==2.2.2
pylint-django==2.0.5
pylint-plugin-utils==0.4
Pyrebase==3.0.27
python-jwt==2.0.1
python-redmine==2.2.1
pytz==2018.9
requests==2.21.0
requests-toolbelt==0.7.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.2
virtualenv==16.4.3
virtualenv-clone==0.5.3
whitenoise==3.3.1
wrapt==1.11.1
